refactor(7141): isolate settlement submission outcomes
Separate transport result classification from the release route, narrow the timeout regression to its conversion boundary, and reuse a test-override predicate so every touched file and function satisfies repository size policy.

Constraint: Preserve the exact durable state and error mapping established by the GREEN checkpoint.

Rejected: Compress the oversized route with denser formatting | extraction keeps responsibilities explicit and reviewable.

Confidence: high

Scope-risk: narrow

Directive: Keep ambiguous and expired settlement persistence in the submission outcome module.

Tested: cargo fmt --all; git diff --check; cargo check for kamn-node and kamn-e2e-harness; strict Clippy for both libraries and binaries; touched Rust size policy GO.

Not-tested: Local test execution remains blocked by the macOS loader stall; the post-refactor Ubuntu fast gate is required.

Related: #7141
